Adds support for bigint. Cleans up index and functions to start from scratch

+++ ndc-duckduckapi/src/generate-config.ts
@@ -18,7 +18,7 @@ if (
 const determineType = (t: string): string => {
   switch (t) {
     case "BIGINT":
-      return "Int";
+      return "BigInt";
     case "BIT":
       return "String";
     case "BOOLEAN":
